give up stopping instance after 10 tries, and copying HOME enviroment variable when launching (needed by Xlib or you get errors like "no protocol specified" etc)
svn path=/trunk/; revision=878

+++ slideshow/daemon.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 
 import multiprocessing, subprocess, sys, threading, traceback
-import time, os.path, re, signal, socket, sqlite3
+import time, os, re, signal, socket, sqlite3
 import dbus, dbus.service
 import cherrypy
 from select import select
@@ -78,7 +78,8 @@ def settings(resolution=None, fullscreen=True):
 	env = dict(
 		DISPLAY=settings['Appearance.Display'],
 		SLIDESHOW_NO_ABORT='',
-		SDL_VIDEO_X11_XRANDR='0'
+		SDL_VIDEO_X11_XRANDR='0',
+		HOME=os.environ['HOME']
 	)
 	for k,v in settings['Env'].items():
 		env['SLIDESHOW_' + k] = v
@@ -181,7 +182,11 @@ class _Daemon(threading.Thread):
 		ipc.Quit()
 		
 		# wait for proper shutdown
+		n = 0
 		while self._instance:
+			if n > 10: # give up
+				break
+
 			try:
 				os.kill(self._instance.pid, 0) # try if process accepts signals
 			except OSError:
@@ -192,6 +197,7 @@ class _Daemon(threading.Thread):
 			
 			print 'waiting for instance to terminate'
 			time.sleep(1)
+			n+=1
 		
 		self._running = False
